+++ catscripts/Government/Countries/Nicaragua/Nicaragua.sh
@@ -1,14 +1,24 @@
 #!/bin/bash
 
-egrep -i 'Nicaragua' newpages.txt >> Nicaragua.txt
+KEYWORDS_NICARAGUA="Nicaragua"
 
-NICARAGUA=$(stat --print=%s Nicaragua.txt)
-
-if [ $NICARAGUA -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Nicaragua.txt"
-  export CATNAME="Nicaragua"
-  $CATEGORIZE
-fi
 
-rm Nicaragua.txt
+  debug_start "Nicaragua"
+
+  NICARAGUA=$(egrep -i "$KEYWORDS_NICARAGUA" newpages.txt)
+
+  if [ "$NICARAGUA" != "" ];
+  then
+    printf "%s" "$NICARAGUA" > Nicaragua.txt
+    export CATFILE="Nicaragua.txt"
+    export CATNAME="Nicaragua"
+    $CATEGORIZE
+    rm Nicaragua.txt
+    unset NICARAGUA
+  fi
+
+  debug_end "Nicaragua"
+
+fi
